GLAN-186: Forms Unit Testing
[1] Form editor javascript
[2] Impact on other units
[3] DRY up javascript code and knock-on effects
[3] BC interface for forms

+++ app/controllers/forms_controller.rb
@@ -52,6 +52,13 @@ class FormsController < ApplicationController
   def edit
     authorize Form
     @form = Form.find(params[:id], params[:namespace])
+    @close_path = history_forms_path(identifier: @form.identifier, scope_id: @form.owner_id)
+    @token = Token.obtain(@form, current_user)
+    ConsoleLogger.debug(C_CLASS_NAME, "edit", "Token=#{@token.to_json}")
+    if @token.nil?
+      flash[:error] = "The item is locked for editing by another user."
+      redirect_to request.referer
+    end
   end
 
   def clone
@@ -61,32 +68,50 @@ class FormsController < ApplicationController
 
   def create
     authorize Form
-    @form = Form.create(params)
+    @form = Form.create_simple(the_params)
     if @form.errors.empty?
       AuditTrail.create_item_event(current_user, @form, "Form created.")
-      render :json => { :data => @form.to_edit}, :status => 200
+      flash[:success] = 'Form was successfully created.'
+      redirect_to forms_path
     else
-      render :json => { :errors => @form.errors.full_messages}, :status => 422
+      flash[:error] = @form.errors.full_messages.to_sentence
+      redirect_to new_form_path
     end
   end
 
   def update
     authorize Form
-    @form = Form.update(params)
-    if @form.errors.empty?
-      AuditTrail.update_item_event(current_user, @form, "Form updated.")
-      render :json => { :data => @form.to_edit}, :status => 200
+    ConsoleLogger.debug(C_CLASS_NAME, "update", "id=#{params[:id]}, namesapce=#{params[:namespace]}")
+    form = Form.find(params[:id], params[:namespace], false)
+    ConsoleLogger.debug(C_CLASS_NAME, "update", "Form=#{form.to_json}")
+    token = Token.find_token(form, current_user)
+    ConsoleLogger.debug(C_CLASS_NAME, "update", "Token=#{token.to_json}")
+    if !token.nil?
+      @form = Form.update(params[:form])
+      ConsoleLogger.debug(C_CLASS_NAME, "update", "Form=#{form.to_json}")
+      if @form.errors.empty?
+        AuditTrail.update_item_event(current_user, @form, "Form updated.") if token.refresh == 1
+        render :json => { :data => @form.to_operation}, :status => 200
+      else
+        render :json => { :errors => @form.errors.full_messages}, :status => 422
+      end
     else
-      render :json => { :errors => @form.errors.full_messages}, :status => 422
+      render :json => {:errors => ["The changes were not saved as the edit lock has timed out."]}, :status => 422
     end
   end
 
   def destroy
     authorize Form
     form = Form.find(params[:id], params[:namespace], false)
-    form.destroy
-    AuditTrail.delete_item_event(current_user, @form, "Form deleted.")
-    redirect_to forms_path
+    token = Token.obtain(form, current_user)
+    if !token.nil?
+      form.destroy
+      AuditTrail.delete_item_event(current_user, form, "Form deleted.")
+      token.release
+    else
+      flash[:error] = "The item is locked for editing by another user."
+    end
+    redirect_to request.referer
   end
 
   def show
